easychef-api/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
djangorestframework-simplejwt==5.2.2
requests~=2.28.2
django-cors-headers~=3.14.0
